requirements.txt
appdirs==1.4.4
apturl==0.5.2
arabic-reshaper==2.1.3
autopep8==1.5.7
bcrypt==3.1.7
beautifulsoup4==4.9.3
blinker==1.4
boto3==1.17.99
botocore==1.20.99
Brlapi==0.7.0
bs4==0.0.1
certifi==2019.11.28
chardet==3.0.4
click==7.1.2
colorama==0.4.3
command-not-found==0.3
cryptography==2.8
cupshelpers==1.0
cycler==0.10.0
dbus-python==1.2.16
defer==1.0.6
Deprecated==1.2.12
distlib==0.3.2
distro==1.4.0
distro-info===0.23ubuntu1
Django==2.2.12
django-crispy-forms==1.12.0
django-filter==2.4.0
django-phone-field==1.8.1
django-storages==1.11.1
dnspython==2.1.0
dominate==2.6.0
duplicity==0.8.12.0
email-validator==1.1.3
entrypoints==0.3
fasteners==0.14.1
filelock==3.0.12
fire==0.4.0
Flask==1.1.4
Flask-Bootstrap==3.3.7.1
flask-buzz==0.1.15
Flask-Cors==3.0.10
Flask-JWT-Extended==4.2.3
Flask-Login==0.5.0
Flask-Mail==0.9.1
flask-marshmallow==0.14.0
flask-praetorian==1.2.0
Flask-SQLAlchemy==2.5.1
Flask-WTF==0.15.1
future==0.18.2
geographiclib==1.52
geopy==2.1.0
greenlet==1.1.0
gunicorn==20.1.0
html5lib==1.0.1
httplib2==0.14.0
idna==2.8
importlib-metadata==1.5.0
inflection==0.3.1
itsdangerous==1.1.0
Jinja2==2.11.3
jmespath==0.10.0
keyring==18.0.1
kiwisolver==1.0.1
language-selector==0.1
launchpadlib==1.10.13
lazr.restfulclient==0.14.2
lazr.uri==1.0.3
lockfile==0.12.2
louis==3.12.0
lxml==4.5.0
macaroonbakery==1.3.1
Mako==1.1.0
MarkupSafe==2.0.1
marshmallow==3.13.0
marshmallow-sqlalchemy==0.26.1
matplotlib==3.1.2
monotonic==1.5
more-itertools==4.2.0
netifaces==0.10.4
numpy==1.21.0
oauthlib==3.1.0
olefile==0.46
pandas==1.2.5
paramiko==2.6.0
passlib==1.7.4
pbr==5.6.0
pdfkit==0.6.1
pendulum==2.1.2
pexpect==4.6.0
Pillow==7.0.0
powerline-status==2.7
prompt-toolkit==1.0.14
protobuf==3.6.1
psutil==5.5.1
py-buzz==1.0.3
pycairo==1.16.2
pycodestyle==2.7.0
pycups==1.9.73
pyfiglet==0.8.post1
Pygments==2.9.0
PyGObject==3.36.0
PyInquirer==1.0.3
PyJWT==2.1.0
pymacaroons==0.13.0
PyNaCl==1.3.0
pyparsing==2.4.6
PyPDF2==1.26.0
pyRFC3339==1.1
PySocks==1.7.1
python-apt==2.0.0+ubuntu0.20.4.5
python-bidi==0.4.2
python-dateutil==2.7.3
python-debian===0.1.36ubuntu1
pytz==2019.3
pytzdata==2020.1
pyxdg==0.26
PyYAML==5.3.1
regex==2021.7.6
reportlab==3.5.34
requests==2.22.0
requests-oauthlib==1.3.0
requests-unixsocket==0.2.0
s3transfer==0.4.2
SecretStorage==2.3.1
selenium==3.141.0
simplejson==3.16.0
six==1.14.0
soupsieve==2.2.1
SQLAlchemy==1.4.22
sqlparse==0.2.4
stevedore==3.3.0
stripe==2.61.0
systemd-python==234
termcolor==1.1.0
testresources==2.0.1
toml==0.10.2
tweepy==3.10.0
ubuntu-advantage-tools==20.3
ubuntu-drivers-common==0.0.0
ufw==0.36
unattended-upgrades==0.1
urllib3==1.25.8
usb-creator==0.3.7
virtualenv==20.4.7
virtualenv-clone==0.5.4
virtualenvwrapper==4.8.4
visitor==0.1.3
wadllib==1.3.3
wcwidth==0.2.5
webencodings==0.5.1
Werkzeug==1.0.1
whitenoise==5.3.0
wrapt==1.12.1
WTForms==2.3.3
xhtml2pdf==0.2.5
xkit==0.0.0
zipp==1.0.0
